+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## Changelog
 (most recent first)
 
+### v0.1 2018-02-16:
+* Changed from examining 10 posts to just the most recent one. The intention is to run this as a cron job every hour (perhaps 2); it's unlikely I'll create a blog post more frequently.
+* Added the creation of a message in my pnut.io dev channel if a new blog post exists.
+* Next step: must create a new pnut.io app to create an independent token. (Currently borrowing the one from `crypto`!)
+
 ### v0.0 2018-02-15:
 * First commit with basic functionality
  * Gets most recent 10 posts from feed.